+++ docs/conf.py
@@ -21,13 +21,13 @@ sys.path.insert(0, os.path.abspath('../quma'))
 # -- Project information -----------------------------------------------------
 
 project = 'quma'
-copyright = '2018, ebene fünf GmbH'
+copyright = '2018-2019, ebene fünf GmbH'
 author = 'ebene fünf GmbH'
 
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '0.1.0a5'
+release = '0.1.0b1'
 
 
 # -- General configuration ---------------------------------------------------